refactor: centralize speech synthesis utility
